refactor billing app colors and theme
