+++ packages/zambdas/src/shared/pdf/pdf-utils.ts
@@ -92,6 +92,13 @@ export function splitLongStringToPageSize(
 
 export const rgbNormalized = (r: number, g: number, b: number): Color => rgb(r / 255, g / 255, b / 255);
 
+export const hexColor = (hex: string): Color => {
+  const value = hex.replace('#', '').slice(0, 6);
+  if (!/^[0-9a-f]{6}$/i.test(value)) throw new Error(`Cannot use '${hex}' as a PDF color`);
+  const channel = (start: number): number => parseInt(value.slice(start, start + 2), 16);
+  return rgbNormalized(channel(0), channel(2), channel(4));
+};
+
 export async function createPdfClient(initialStyles: PdfClientStyles): Promise<PdfClient> {
   const pdfDoc = await PDFDocument.create();
   pdfDoc.registerFontkit(fontkit);
